Create an AWSMock class
Mocking AWS has been kinda painful, mostly because I use the
`.promise()` pattern. This means method chaining ...

where I expect parameters to be passed to the first method, but return
values to be wrapped in a Promise as the return of the chained method.

This could be worth open sourcing, especially if I manage to pull in the
parameter validation code from the AWS SDK itself.